requirements.txt
asgiref==3.8.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
defusedxml==0.7.1
Django==5.2.8
django-bootstrap5==25.2
fonttools==4.59.2
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytils==0.4.1
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
